% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_app.R
\name{search_app_apple}
\alias{search_app_apple}
\title{Search in App Store}
\usage{
search_app_apple(term, country = "es", platform = "software")
}
\arguments{
\item{term}{word(s) to search apps about}

\item{country}{market where the user desired to search the app}

\item{platform}{any, iPhone, iPad or iMac}
}
\value{
tibble with the apps found for the search
}
\description{
Return the list obtaining for searching the text giving as parameter
}
