% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_app.R
\name{search_app_google}
\alias{search_app_google}
\title{Search in Google Play}
\usage{
search_app_google(
  term,
  num = 20,
  lang = "en",
  country = "us",
  fullDetail = FALSE,
  price = "all"
)
}
\arguments{
\item{term}{the term to search by}

\item{num}{optional the amount of apps to retrieve. defaults to 20, max is 250):}

\item{lang}{(optional, defaults to 'en'): the two letter language code used to retrieve the applications.}

\item{country}{(optional, defaults to 'us'): the two letter country code used to retrieve the applications.}

\item{fullDetail}{(optional, defaults to false): if true, an extra request will be made for every resulting app to fetch its full detail.}

\item{price}{(optional, defaults to all): allows to control if the results apps are free, paid or both.
\itemize{
 \item{"all": }{Free and paid}
 \item{"free": }{Free apps only}
 \item{"paid": }{Paid apps only}
}}
}
\value{
tibble containing the information of the apps retrieved in the search
\itemize{
\item \code{title} - Full name of the app in Google Play
\item description
\item descriptionHTML
\item summary
\item installs
\item minInstalls
\item maxInstalls
\item score
\item scoreText
\item ratings
\item reviews
\item histogram
\item price
\item free
\item currency
\item priceText
\item offersIAP
\item IAPRange
\item size
\item androidVersion
\item androidVersionText
\item developer
\item developerId
\item developerEmail
\item developerWebsite
\item developerAddress
\item privacyPolicy
\item developerInternalID
\item genre
\item genreId
\item familyGenre
\item familyGenreId
\item icon
\item headerImage
\item screenshots
\item video
\item videoImage
\item contentRating
\item contentRatingDescription
\item adSupported
\item released
\item updated
\item version
\item recentChanges
\item comments
\item editorsChoice
\item appId
\item url
}
}
\description{
Retrieves a list of apps in Google Play that results of searching by the
given term.
}
